# bff/src/main/java/com/globallogic/bff/dto/ProductDTO.java
package com.globallogic.bff.dto;

import java.io.Serializable;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ProductDTO implements Serializable {

    private static final long serialVersionUID = 7327026715164893273L;

    private String id;
    private String description;
    private int stock;
    private double price;
}