# bff/src/main/java/com/globallogic/bff/dto/ProductReportDTO.java
package com.globallogic.bff.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ProductReportDTO {

    private String id;
    private String description;
    private int totalUnitsSold;


}