# bff/src/main/java/com/globallogic/bff/dto/SoldProductsReportDTO.java
package com.globallogic.bff.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class SoldProductsReportDTO {

    private List<ProductReportDTO> products;
}