# tarjetas/src/main/java/com/globallogic/javaneta/cards/model/dto/CardsResponseDTO.java
package com.globallogic.javaneta.cards.model.dto;

public class CardsResponseDTO extends CardsDTO {

    public CardsResponseDTO(String number, String clientId, String company) {
        super(number, clientId, company);
    }

}